Add base interface
Add the methods that Melophony will use to get the available
providers.
For the ITrackProvider interface:
Add the method to let Melophony call the provider to get the fields it
will need to add the track.
Add the method that will effectively add the track file.
Add the callback that the provider should call to give track
 information back to Melophony.